remove writer_go18*
